requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --config=pyproject.toml requirements/requirements.in
#
aiohttp==3.8.5
    # via openai
aiosignal==1.3.1
    # via aiohttp
amqp==5.1.1
    # via kombu
anyascii==0.3.2
    # via wagtail
asgiref==3.7.2
    # via
    #   channels
    #   channels-redis
    #   django
async-timeout==4.0.3
    # via
    #   aiohttp
    #   redis
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
beautifulsoup4==4.11.2
    # via wagtail
billiard==4.1.0
    # via celery
boto3==1.28.38
    # via -r requirements/requirements.in
botocore==1.31.38
    # via
    #   boto3
    #   s3transfer
brotli==1.0.9
    # via whitenoise
celery[redis]==5.3.4
    # via -r requirements/requirements.in
celery-progress==0.3
    # via -r requirements/requirements.in
certifi==2023.7.22
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
channels==4.0.0
    # via
    #   -r requirements/requirements.in
    #   channels-redis
channels-redis==4.1.0
    # via -r requirements/requirements.in
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cryptography==41.0.3
    # via
    #   django-anymail
    #   pyjwt
defusedxml==0.7.1
    # via
    #   python3-openid
    #   willow
dj-stripe==2.8.1
    # via -r requirements/requirements.in
django==4.2.6
    # via
    #   -r requirements/requirements.in
    #   channels
    #   dj-stripe
    #   django-allauth
    #   django-allauth-2fa
    #   django-anymail
    #   django-filter
    #   django-modelcluster
    #   django-otp
    #   django-permissionedforms
    #   django-storages
    #   django-taggit
    #   django-treebeard
    #   django-waffle
    #   djangorestframework
    #   drf-spectacular
    #   wagtail
django-allauth==0.57.0
    # via
    #   -r requirements/requirements.in
    #   django-allauth-2fa
django-allauth-2fa==0.11.1
    # via -r requirements/requirements.in
django-anymail[mailgun]==10.1
    # via -r requirements/requirements.in
django-environ==0.11.2
    # via -r requirements/requirements.in
django-filter==23.2
    # via wagtail
django-modelcluster==6.0
    # via wagtail
django-otp==1.2.2
    # via django-allauth-2fa
django-permissionedforms==0.1
    # via wagtail
django-storages==1.13.2
    # via -r requirements/requirements.in
django-taggit==4.0.0
    # via wagtail
django-treebeard==4.7
    # via wagtail
django-waffle==4.0.0
    # via -r requirements/requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements/requirements.in
    #   drf-spectacular
    #   wagtail
docutils==0.20.1
    # via -r requirements/requirements.in
draftjs-exporter==2.1.7
    # via wagtail
drf-spectacular==0.26.4
    # via -r requirements/requirements.in
et-xmlfile==1.1.0
    # via openpyxl
filetype==1.2.0
    # via willow
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
html5lib==1.1
    # via wagtail
idna==3.4
    # via
    #   requests
    #   yarl
inflection==0.5.1
    # via drf-spectacular
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.19.0
    # via drf-spectacular
jsonschema-specifications==2023.7.1
    # via jsonschema
kombu==5.3.2
    # via celery
l18n==2021.3
    # via wagtail
markdown==3.4.4
    # via -r requirements/requirements.in
msgpack==1.0.6
    # via channels-redis
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
oauthlib==3.2.2
    # via requests-oauthlib
openai==0.27.10
    # via -r requirements/requirements.in
openpyxl==3.1.2
    # via wagtail
pillow==10.0.0
    # via
    #   pillow-heif
    #   wagtail
pillow-heif==0.13.0
    # via willow
prompt-toolkit==3.0.39
    # via click-repl
psycopg2-binary==2.9.7
    # via -r requirements/requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via django-allauth
pypng==0.20220715.0
    # via qrcode
python-dateutil==2.8.2
    # via
    #   botocore
    #   celery
python3-openid==3.2.0
    # via django-allauth
pytz==2023.3
    # via
    #   django-modelcluster
    #   djangorestframework
    #   l18n
pyyaml==6.0.1
    # via drf-spectacular
qrcode==7.4.2
    # via django-allauth-2fa
redis==4.6.0
    # via
    #   celery
    #   channels-redis
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   -r requirements/requirements.in
    #   django-allauth
    #   django-anymail
    #   openai
    #   requests-oauthlib
    #   stripe
    #   wagtail
requests-oauthlib==1.3.1
    # via django-allauth
rpds-py==0.10.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.6.2
    # via boto3
sentry-sdk==1.30.0
    # via -r requirements/requirements.in
six==1.16.0
    # via
    #   html5lib
    #   l18n
    #   python-dateutil
soupsieve==2.4.1
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
stripe==4.2.0
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
telepath==0.3.1
    # via wagtail
tqdm==4.66.1
    # via openai
typing-extensions==4.7.1
    # via qrcode
tzdata==2023.3
    # via celery
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.16
    # via
    #   botocore
    #   django-anymail
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wagtail==5.1.1
    # via -r requirements/requirements.in
wcwidth==0.2.6
    # via prompt-toolkit
webencodings==0.5.1
    # via html5lib
whitenoise[brotli]==6.5.0
    # via -r requirements/requirements.in
willow[heif]==1.6.1
    # via wagtail
yarl==1.9.2
    # via aiohttp
